feat: Add: 1.MniIO SyncMinioTemplate API;2.Update README.
