+++ spring-boot-minio-autoconfigure/src/main/java/io/github/photowey/minio/spring/boot/autoconfigure/template/MinioTemplate.java
@@ -34,8 +34,6 @@ import java.util.concurrent.TimeUnit;
  */
 public interface MinioTemplate {
 
-    String MINIO_TEMPLATE_BEAN_NAME = "minioTemplate";
-
     String DEFAULT_STREAM_CONTENT_TYPE = "application/octet-stream";
 
     // ----------------------------------------------------------------
